# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: User Microservice
# during dev, should point to your local machine
host: localhost:3000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/x-www-form-urlencoded
  - multipart/form-data
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /users/{id}:
    # our controller name
    x-swagger-router-controller: users
    get:
      operationId: getOne
      description: get a user
      # define the success and error response
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: update
      description: update a user
      # define the parameters
      parameters:
        - name: id
          description: user id
          type: string
          in: path
          required: true
        - name: firstName
          description: user firstName
          in: body
          required: true
          schema:
            $ref: "#/definitions/Users"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteUser
      description: delete a user
      # define the parameters
      parameters:
        - name: id
          description: user id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  Users:
    type: object
    properties:
      firstName:
        type: string
        description: users first name
      lastName:
        type: string
        description: users last name
      login:
        type: string
        description: users email address and account login
      password:
        type: string
        description: users account password
    required:
      - firstName
      - lastName
      - login
      - password
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  GetUserResponse:
      required:
        - _id
        - firstName
        - lastName
        - login
        - password
      properties:
        _id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        login:
          type: string
        password:
          type: string
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short general response comment
    required:
      - success
      - description